+++ vaden/lib/src/types/param_parse.dart
@@ -1,0 +1,7 @@
+abstract class ParamParse<P, R> {
+  const ParamParse();
+
+  R toJson(P param);
+
+  P fromJson(R json);
+}