+++ vaden/lib/src/types/types.dart
@@ -1,5 +1,6 @@
 export 'disposable.dart';
 export 'guard.dart';
 export 'middleware.dart';
+export 'param_parse.dart';
 export 'response_exception.dart';
 export 'validator.dart';