+++ vaden/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.7 (2025/03/28)
+- Fix abstract Controller methods.
+- Fix Complex object parse.
+
 ## 0.0.6
 
 - Start ALPHA